+++ src/extension.ts
@@ -3,7 +3,6 @@
 import * as vscode from "vscode";
 import AddClass from "./commands/add-class";
 import AddInterface from "./commands/add-interface";
-import * as path from "path";
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed