+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # DotNet Core Add Changelog
 
+## [0.0.2] - 2020-05-26
+
+### Refactored
+
+- Refactored code for calculate the namespace
+- Fix error wrong namespace when `RootNamespace` node is present on `csproj`
+
 ## [0.0.1] - 2020-05-22
 
 ### Added